Moved gigs into lastfm.py
